+++ src/rpc.rs
@@ -8,7 +8,7 @@ use std::{
 
 use anyhow::anyhow;
 use client::{
-    blobs::{BlobInfo, BlobStatus, IncompleteBlobInfo, MemClient, WrapOption},
+    blobs::{self, BlobInfo, BlobStatus, DownloadMode, IncompleteBlobInfo, MemClient, WrapOption},
     tags::TagInfo,
     MemConnector,
 };
@@ -16,6 +16,7 @@ use futures_buffered::BufferedStreamExt;
 use futures_lite::StreamExt;
 use futures_util::{FutureExt, Stream};
 use genawaiter::sync::{Co, Gen};
+use iroh::{Endpoint, NodeAddr};
 use iroh_base::hash::{BlobFormat, HashAndFormat};
 use iroh_io::AsyncSliceReader;
 use proto::{
@@ -39,15 +40,21 @@ use quic_rpc::{
     RpcClient, RpcServer,
 };
 use tokio_util::task::AbortOnDropHandle;
+use tracing::{debug, warn};
 
 use crate::{
+    downloader::{DownloadRequest, Downloader},
     export::ExportProgress,
     format::collection::Collection,
-    get::db::DownloadProgress,
-    net_protocol::{BlobDownloadRequest, Blobs},
+    get::{
+        db::{DownloadProgress, GetState},
+        Stats,
+    },
+    net_protocol::{BlobDownloadRequest, Blobs, BlobsInner},
     provider::{AddProgress, BatchAddPathProgress},
     store::{ConsistencyCheckProgress, ImportProgress, MapEntry, ValidateProgress},
     util::{
+        local_pool::LocalPoolHandle,
         progress::{AsyncChannelProgressSender, ProgressSender},
         SetTagOption,
     },
@@ -62,10 +69,62 @@ const RPC_BLOB_GET_CHUNK_SIZE: usize = 1024 * 64;
 const RPC_BLOB_GET_CHANNEL_CAP: usize = 2;
 
 impl<D: crate::store::Store> Blobs<D> {
-    /// Spawns an in-memory RPC client and server pair.
-    #[must_use = "Dropping the RpcHandler will stop the client"]
-    pub fn spawn_rpc(&self) -> RpcHandler {
-        RpcHandler::new(self)
+    /// Get a client for the blobs protocol
+    pub fn client(&self) -> &blobs::MemClient {
+        &self
+            .rpc_handler
+            .get_or_init(|| RpcHandler::new(&self.inner))
+            .client
+    }
+
+    /// Handle an RPC request
+    pub async fn handle_rpc_request<C>(
+        self,
+        msg: Request,
+        chan: RpcChannel<RpcService, C>,
+    ) -> std::result::Result<(), RpcServerError<C>>
+    where
+        C: ChannelTypes<RpcService>,
+    {
+        Handler(self.inner.clone())
+            .handle_rpc_request(msg, chan)
+            .await
+    }
+}
+
+#[derive(Clone)]
+struct Handler<S>(Arc<BlobsInner<S>>);
+
+impl<S> Deref for Handler<S> {
+    type Target = BlobsInner<S>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<D: crate::store::Store> Handler<D> {
+    fn store(&self) -> &D {
+        &self.0.store
+    }
+
+    fn rt(&self) -> &LocalPoolHandle {
+        &self.0.rt
+    }
+
+    fn endpoint(&self) -> &Endpoint {
+        &self.0.endpoint
+    }
+
+    fn downloader(&self) -> &Downloader {
+        &self.0.downloader
+    }
+
+    #[cfg(feature = "rpc")]
+    pub(crate) async fn batches(
+        &self,
+    ) -> tokio::sync::MutexGuard<'_, crate::net_protocol::BlobBatches> {
+        self.0.batches.lock().await
     }
 
     /// Handle an RPC request
@@ -869,6 +928,142 @@ impl<D: crate::store::Store> Blobs<D> {
 
         Ok(CreateCollectionResponse { hash, tag })
     }
+
+    pub(crate) async fn download(
+        &self,
+        endpoint: Endpoint,
+        req: BlobDownloadRequest,
+        progress: AsyncChannelProgressSender<DownloadProgress>,
+    ) -> anyhow::Result<()> {
+        let BlobDownloadRequest {
+            hash,
+            format,
+            nodes,
+            tag,
+            mode,
+        } = req;
+        let hash_and_format = HashAndFormat { hash, format };
+        let temp_tag = self.store().temp_tag(hash_and_format);
+        let stats = match mode {
+            DownloadMode::Queued => {
+                self.download_queued(endpoint, hash_and_format, nodes, progress.clone())
+                    .await?
+            }
+            DownloadMode::Direct => {
+                self.download_direct_from_nodes(endpoint, hash_and_format, nodes, progress.clone())
+                    .await?
+            }
+        };
+
+        progress.send(DownloadProgress::AllDone(stats)).await.ok();
+        match tag {
+            SetTagOption::Named(tag) => {
+                self.store().set_tag(tag, Some(hash_and_format)).await?;
+            }
+            SetTagOption::Auto => {
+                self.store().create_tag(hash_and_format).await?;
+            }
+        }
+        drop(temp_tag);
+
+        Ok(())
+    }
+
+    async fn download_queued(
+        &self,
+        endpoint: Endpoint,
+        hash_and_format: HashAndFormat,
+        nodes: Vec<NodeAddr>,
+        progress: AsyncChannelProgressSender<DownloadProgress>,
+    ) -> anyhow::Result<Stats> {
+        /// Name used for logging when new node addresses are added from gossip.
+        const BLOB_DOWNLOAD_SOURCE_NAME: &str = "blob_download";
+
+        let mut node_ids = Vec::with_capacity(nodes.len());
+        let mut any_added = false;
+        for node in nodes {
+            node_ids.push(node.node_id);
+            if !node.info.is_empty() {
+                endpoint.add_node_addr_with_source(node, BLOB_DOWNLOAD_SOURCE_NAME)?;
+                any_added = true;
+            }
+        }
+        let can_download = !node_ids.is_empty() && (any_added || endpoint.discovery().is_some());
+        anyhow::ensure!(can_download, "no way to reach a node for download");
+        let req = DownloadRequest::new(hash_and_format, node_ids).progress_sender(progress);
+        let handle = self.downloader().queue(req).await;
+        let stats = handle.await?;
+        Ok(stats)
+    }
+
+    #[tracing::instrument("download_direct", skip_all, fields(hash=%hash_and_format.hash.fmt_short()))]
+    async fn download_direct_from_nodes(
+        &self,
+        endpoint: Endpoint,
+        hash_and_format: HashAndFormat,
+        nodes: Vec<NodeAddr>,
+        progress: AsyncChannelProgressSender<DownloadProgress>,
+    ) -> anyhow::Result<Stats> {
+        let mut last_err = None;
+        let mut remaining_nodes = nodes.len();
+        let mut nodes_iter = nodes.into_iter();
+        'outer: loop {
+            match crate::get::db::get_to_db_in_steps(
+                self.store().clone(),
+                hash_and_format,
+                progress.clone(),
+            )
+            .await?
+            {
+                GetState::Complete(stats) => return Ok(stats),
+                GetState::NeedsConn(needs_conn) => {
+                    let (conn, node_id) = 'inner: loop {
+                        match nodes_iter.next() {
+                            None => break 'outer,
+                            Some(node) => {
+                                remaining_nodes -= 1;
+                                let node_id = node.node_id;
+                                if node_id == endpoint.node_id() {
+                                    debug!(
+                                        ?remaining_nodes,
+                                        "skip node {} (it is the node id of ourselves)",
+                                        node_id.fmt_short()
+                                    );
+                                    continue 'inner;
+                                }
+                                match endpoint.connect(node, crate::protocol::ALPN).await {
+                                    Ok(conn) => break 'inner (conn, node_id),
+                                    Err(err) => {
+                                        debug!(
+                                            ?remaining_nodes,
+                                            "failed to connect to {}: {err}",
+                                            node_id.fmt_short()
+                                        );
+                                        continue 'inner;
+                                    }
+                                }
+                            }
+                        }
+                    };
+                    match needs_conn.proceed(conn).await {
+                        Ok(stats) => return Ok(stats),
+                        Err(err) => {
+                            warn!(
+                                ?remaining_nodes,
+                                "failed to download from {}: {err}",
+                                node_id.fmt_short()
+                            );
+                            last_err = Some(err);
+                        }
+                    }
+                }
+            }
+        }
+        match last_err {
+            Some(err) => Err(err.into()),
+            None => Err(anyhow!("No nodes to download from provided")),
+        }
+    }
 }
 
 /// An in memory rpc handler for the blobs rpc protocol
@@ -880,7 +1075,7 @@ impl<D: crate::store::Store> Blobs<D> {
 /// where it will be kept alive. This struct will capture a copy of
 /// [`crate::net_protocol::Blobs`] and keep it alive.
 #[derive(Debug)]
-pub struct RpcHandler {
+pub(crate) struct RpcHandler {
     /// Client to hand out
     client: MemClient,
     /// Handler task
@@ -896,19 +1091,15 @@ impl Deref for RpcHandler {
 }
 
 impl RpcHandler {
-    fn new<D: crate::store::Store>(blobs: &Blobs<D>) -> Self {
+    fn new<D: crate::store::Store>(blobs: &Arc<BlobsInner<D>>) -> Self {
         let blobs = blobs.clone();
         let (listener, connector) = quic_rpc::transport::flume::channel(1);
         let listener = RpcServer::new(listener);
         let client = RpcClient::new(connector);
         let client = MemClient::new(client);
-        let _handler = listener
-            .spawn_accept_loop(move |req, chan| blobs.clone().handle_rpc_request(req, chan));
+        let _handler = listener.spawn_accept_loop(move |req, chan| {
+            Handler(blobs.clone()).handle_rpc_request(req, chan)
+        });
         Self { client, _handler }
-    }
-
-    /// Get a reference to the rpc client api
-    pub fn client(&self) -> &MemClient {
-        &self.client
     }
 }
