+++ tests/blobs.rs
@@ -13,7 +13,7 @@ async fn blobs_gc_smoke() -> TestResult<()> {
     let pool = LocalPool::default();
     let endpoint = Endpoint::builder().bind().await?;
     let blobs = Blobs::memory().build(pool.handle(), &endpoint);
-    let client = blobs.spawn_rpc();
+    let client = blobs.client();
     blobs.start_gc(GcConfig {
         period: Duration::from_millis(1),
         done_callback: None,
@@ -32,7 +32,7 @@ async fn blobs_gc_protected() -> TestResult<()> {
     let pool = LocalPool::default();
     let endpoint = Endpoint::builder().bind().await?;
     let blobs = Blobs::memory().build(pool.handle(), &endpoint);
-    let client = blobs.spawn_rpc();
+    let client = blobs.client();
     let h1 = client.add_bytes(b"test".to_vec()).await?;
     let protected = Arc::new(Mutex::new(Vec::new()));
     blobs.add_protected(Box::new({
